requirements.txt
streamlit
pandas
plotly
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
openpyxl